feat: create update test case controller

+++ web-infras/parser/tests/parserRunner/updateController/controller.ts
@@ -1,15 +1,15 @@
 import { ChildProcess, fork } from "node:child_process";
 import path from "node:path";
-import { getTestSuiteIterator } from "./getTestCase";
+import { getTestSuiteIterator } from "../helpers/getTestCase";
 import {
   TestCase,
   TestCaseIterator,
-  TestCaseResult,
-  PassTestCaseResult,
-  FailedTestCasesResult,
-  SkipTestCaseResult,
-  TimeoutTestCaseResult,
-} from "./type";
+  UpdateTestCaseResult,
+  UpdatePassTestCaseResult,
+  UpdateFailedTestCaseResult,
+  UpdateTimeoutTestCaseResult,
+  UpdateTestCaseResultSuite,
+} from "../type";
 
 const WORKER_NUM = 5;
 
@@ -17,18 +17,15 @@ async function createChildPorcessController(
   iter: TestCaseIterator,
   processes: Array<ChildProcess>,
   index: number,
-  passResult: Array<PassTestCaseResult>,
-  failedResult: Array<FailedTestCasesResult>,
-  skipResult: Array<SkipTestCaseResult>,
-  timeoutResult: Array<TimeoutTestCaseResult>,
+  suite: UpdateTestCaseResultSuite,
 ) {
   let childProcess: ChildProcess = processes[index];
-  async function waitForTestCaseResult(testCase: TestCase): Promise<TestCaseResult> {
+  async function waitForTestCaseResult(testCase: TestCase): Promise<UpdateTestCaseResult> {
     return new Promise((resolve) => {
       childProcess.send(testCase);
       let flag = false;
       let timerId: NodeJS.Timeout | null = null;
-      childProcess.on("message", (testCaseResult: TestCaseResult) => {
+      childProcess.on("message", (testCaseResult: UpdateTestCaseResult) => {
         flag = true;
         if (timerId) clearTimeout(timerId);
         resolve(testCaseResult);
@@ -54,61 +51,50 @@ async function createChildPorcessController(
       childProcess.kill();
       break;
     }
+    if (testCase.kind === "ExpectFailedTestCase") {
+      continue;
+    }
     const result = await waitForTestCaseResult(testCase);
     childProcess.removeAllListeners();
     switch (result.kind) {
       case "Timeout": {
-        console.log("Timeout", testCase);
-        timeoutResult.push(result);
+        suite.timeoutResult.push(result);
         childProcess.kill();
         processes[index] = fork(path.join(__dirname, "worker.ts"));
         childProcess = processes[index];
         break;
       }
-      case "Skip": {
-        skipResult.push(result);
-        break;
-      }
       case "ExpectPass":
-      case "ExpectFailed": {
-        passResult.push(result);
+      case "ExpectFailed" /** expect failed should be unreach */: {
+        suite.passResult.push(result);
         break;
       }
       default: {
-        failedResult.push(result);
+        suite.failedResult.push(result);
         break;
       }
     }
   }
 }
 
-export async function startController() {
+export async function startUpdateController() {
   const testCaseIter = await getTestSuiteIterator();
   const processes: Array<ChildProcess> = [];
   for (let i = 0; i < WORKER_NUM; ++i) {
     processes.push(fork(path.join(__dirname, "worker.ts")));
   }
-  const passResult: Array<PassTestCaseResult> = [];
-  const failedResult: Array<FailedTestCasesResult> = [];
-  const skipResult: Array<SkipTestCaseResult> = [];
-  const timeoutResult: Array<TimeoutTestCaseResult> = [];
+  const passResult: Array<UpdatePassTestCaseResult> = [];
+  const failedResult: Array<UpdateFailedTestCaseResult> = [];
+  const timeoutResult: Array<UpdateTimeoutTestCaseResult> = [];
+  const suite: UpdateTestCaseResultSuite = {
+    passResult,
+    failedResult,
+    timeoutResult,
+  };
   const promises = processes.map((_, index) =>
-    createChildPorcessController(
-      testCaseIter,
-      processes,
-      index,
-      passResult,
-      failedResult,
-      skipResult,
-      timeoutResult,
-    ),
+    createChildPorcessController(testCaseIter, processes, index, suite),
   );
   await Promise.all(promises);
 
-  return {
-    passResult,
-    failedResult,
-    skipResult,
-    timeoutResult,
-  };
+  return suite;
 }
